react: refactor tests
