+++ src/ui/react/mocks.ts
@@ -1,4 +1,24 @@
+import { ExecutionContext } from "ava";
 import React from "react";
+import { configure } from "./configure";
+import { Data } from "../../core";
+import { makeDummyAdapter, makeWindowMock } from "../mocks";
+
+export function build<D extends Data<any>>(t: ExecutionContext, data: D) {
+  type Domain = D extends Data<infer X> ? X : never;
+  const react = makeReactMock();
+  const configured = configure<Domain>({
+    data: data,
+    React: react
+  });
+  const { db } = configured.create({
+    network: makeDummyAdapter(data),
+    onError: err => t.fail(err.message),
+    window: makeWindowMock()
+  });
+  react._setContextValue(db);
+  return configured;
+}
 
 export function makeReactMock() {
   let contextValue = "missing-provider";
@@ -17,5 +37,3 @@ export function makeReactMock() {
     }
   };
 }
-
-export { makeWindowMock, networkMock } from "../mocks";
