--- src/ui/react/use-multiple-data.test.ts
@@ -1,33 +1,0 @@
-import test from "ava";
-
-import { configure } from "./configure";
-import { makeReactMock, makeWindowMock, networkMock } from "./mocks";
-
-test("useMultipleData sanity", t => {
-  const react = makeReactMock();
-  const { create, useMultipleData } = configure({
-    data: {
-      tasks: {
-        "1": { revision: "1", value: "Buy milk" },
-        "2": { revision: "1", value: "Fix roof" }
-      },
-      projects: {
-        "1": { revision: "1", value: { size: "big" } }
-      },
-      unused: { "0": { revision: "1", value: "Unused" } }
-    },
-    React: react
-  });
-  const { db } = create({
-    network: networkMock,
-    onError: err => t.fail(err.message),
-    window: makeWindowMock()
-  });
-  react._setContextValue(db);
-
-  const [tasks, updateTask] = useMultipleData("tasks", ["1", "2"]);
-  if (tasks === "loading") return t.fail("Expected tasks to not be loading");
-  t.is(tasks["1"], "Buy milk");
-  t.is(tasks["2"], "Fix roof");
-  t.is(typeof updateTask, "function");
-});
